Consolidate Citation types
