+++ src/shared/components/literature-citations/__mocks__/literatureCitationData.ts
@@ -1,7 +1,7 @@
-import { LiteratureResultsAPI } from '../../../../uniprotkb/types/literatureTypes';
+import { CitationsAPIModel } from '../../../../supporting-data/citations/adapters/citationsConverter';
 
 // uniprot/api/citations/14702039
-const mock: LiteratureResultsAPI = {
+const mock: CitationsAPIModel = {
   citation: {
     citationType: 'UniProt indexed literatures',
     authors: [
