+++ src/shared/types/apiModel.ts
@@ -32,19 +32,3 @@ export type Xref = {
   isoformId?: string;
   implicit?: true;
 };
-
-export type Citation = {
-  citationType?: string;
-  authors?: string[];
-  citationCrossReferences?: Xref[];
-  title?: string;
-  publicationDate?: string;
-  journal?: string;
-  firstPage?: string;
-  lastPage?: string;
-  volume?: string;
-  completeAuthorList?: boolean;
-  literatureAbstract?: string;
-  authoringGroup?: string[];
-  submissionDatabase?: string;
-};
